refactor: validations for uuid

+++ backend/src/app/controllers/ContractController.js
@@ -41,18 +41,27 @@ class ContractController {
       name, email, phone, category_id,
     } = req.body;
 
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Category not found, invalid id' });
+    }
+
     if (!name) {
       return res.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
 
-    if (contactExists) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+      if (contactExists) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     return res.status(201).json(contact);
@@ -72,27 +81,34 @@ class ContractController {
       return res.status(400).json({ error: 'Contact not found, invalid id' });
     }
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return res.status(404).json({ error: 'Contact not found' });
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Category not found, invalid id' });
     }
 
     if (!name) {
       return res.status(400).json({ error: 'Name is required' });
     }
 
-    const contactEmailExists = await ContactsRepository.findByEmail(email);
-    if (contactEmailExists && contactEmailExists.id !== id) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+    const contactExists = await ContactsRepository.findById(id);
+    if (!contactExists) {
+      return res.status(404).json({ error: 'Contact not found' });
+    }
+
+    if (email) {
+      const contactEmailExists = await ContactsRepository.findByEmail(email);
+
+      if (contactEmailExists && contactEmailExists.id !== id) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(
       id,
       {
         name,
-        email,
+        email: email || null,
         phone,
-        category_id,
+        category_id: category_id || null,
       },
     );
 
@@ -104,6 +120,10 @@ class ContractController {
    */
   async delete(req, res) {
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Contact not found, invalid id' });
+    }
 
     await ContactsRepository.delete(id);
